updated question architecture & hid db creds
