+++ .gitignore
@@ -38,3 +38,6 @@ jspm_packages
 
 # Optional REPL history
 .node_repl_history
+
+#credentials
+credentials.js
