+++ package.json
@@ -20,26 +20,28 @@
   "homepage": "https://github.com/Greenfield-project-team-Io/questionsQueue#readme",
   "dependencies": {
     "babel-core": "^6.24.1",
-    "react": "^15.5.4",
-    "react-dom": "^15.5.4",
-    "mongodb": "^2.2.26",
-    "mongoose": "^4.9.6",
     "body-parser": "^1.17.1",
     "express": "^4.15.2",
-    "http": "0.0.0"
+    "http": "0.0.0",
+    "mongodb": "^2.2.26",
+    "mongoose": "^4.9.6",
+    "morgan": "^1.8.1",
+    "path": "^0.12.7",
+    "react": "^15.5.4",
+    "react-dom": "^15.5.4"
   },
   "devDependencies": {
     "babel-eslint": "^7.2.3",
     "babel-loader": "^7.0.0",
+    "babel-preset-es2015": "^6.24.1",
+    "babel-preset-react": "^6.24.1",
     "eslint": "^3.19.0",
     "eslint-config-airbnb": "^14.1.0",
     "eslint-loader": "^1.7.1",
     "eslint-plugin-import": "^2.2.0",
     "eslint-plugin-jsx-a11y": "^4.0.0",
     "eslint-plugin-react": "^6.10.3",
-    "babel-loader": "^7.0.0",
-    "babel-preset-es2015": "^6.24.1",
-    "babel-preset-react": "^6.24.1",
+    "nodemon": "^1.11.0",
     "webpack": "^2.4.1"
   }
 }
