+++ src/db/db-schema.js
@@ -5,6 +5,7 @@ var questionSchema = new mongoose.Schema({
 	questionText : String,
 	votes : Number,
 	answered : Boolean,
+	// timestamps: { createdAt: 'created_at' },
 });
 
 //exporting the questionQueue schema with questionQueue instance
